+++ lib/api_request.dart
@@ -8,3 +8,4 @@ export 'src/api_request_options.dart';
 export 'src/interceptors/api_interceptor.dart';
 export 'src/performance/api_request_performance.dart';
 export 'src/performance/performance_report.dart';
+export 'src/simple_api_request.dart';