+++ lib/src/actions/request_action.dart
@@ -5,13 +5,12 @@ import 'package:dartz/dartz.dart';
 import 'package:dio/dio.dart';
 
 import '../api_request_error.dart';
-import '../api_request_exception.dart';
 import '../utils/api_request_utils.dart';
 
 enum RequestMethod { GET, POST, PUT, DELETE }
 
 typedef ResponseBuilder<T> = T Function(dynamic);
-typedef ErrorHandler = Function(ApiRequestError error);
+typedef ErrorHandler = Function(ActionRequestError error);
 typedef SuccessHandler<T> = Function(T? response);
 
 abstract class RequestAction<T, R extends ApiRequest> {
@@ -53,8 +52,9 @@ abstract class RequestAction<T, R extends ApiRequest> {
 
   ErrorHandler onError = (error) => {};
   SuccessHandler<T> onSuccess = (response) => {};
+  Function onDone = () => {};
 
-  void _streamError(ApiRequestError error) {
+  void _streamError(ActionRequestError error) {
     this.onError(error);
     if (ApiRequestOptions.instance!.onError != null) {
       ApiRequestOptions.instance!.onError!(error);
@@ -81,57 +81,51 @@ abstract class RequestAction<T, R extends ApiRequest> {
     return this;
   }
 
-  Future<Either<ApiRequestException, T?>> run() async {
-    Response? response;
-    Either<ApiRequestException, T?>? toReturn;
-    await _execute().then((value) {
-      response = value;
-    }).catchError((e) {
-      toReturn = left(ApiRequestException(
-          message: e.toString(), type: ApiExceptionType.custom));
-    });
-
-    if (response != null) {
-      int statusCode = response?.statusCode ?? 0;
-      if (statusCode >= 200 && statusCode < 300) {
-        toReturn = right(responseBuilder(response?.data));
-      }
-      if (statusCode >= 400 && statusCode < 500) {
-        toReturn = left(ApiRequestException(
-            message: response?.data?["message"]?.toString() ??
-                response?.statusMessage ??
-                "client_error",
-            type: ApiExceptionType.client,
-            errors: response?.data?["errors"],
-            statusCode: response?.statusCode,
-            statusMessage: response?.statusMessage));
-      }
-      if (statusCode >= 500) {
-        toReturn = left(ApiRequestException(
-            message: "${response?.statusCode} : ${response?.statusMessage}",
-            type: ApiExceptionType.server,
-            statusCode: statusCode,
-            statusMessage: response?.statusMessage));
-      }
-    } else {
-      toReturn = left(
-          ApiRequestException(message: "error", type: ApiExceptionType.custom));
+  RequestAction listen(
+      {Function? onStart,
+      Function? onDone,
+      SuccessHandler<T>? onSuccess,
+      ErrorHandler? onError}) {
+    if (onStart != null) {
+      this.onStart = onStart;
     }
-
-    return toReturn!;
+    if (onDone != null) {
+      this.onDone = onDone;
+    }
+    if (onSuccess != null) {
+      this.onSuccess = onSuccess;
+    }
+    if (onError != null) {
+      this.onError = onError;
+    }
+    return this;
   }
 
-  @Deprecated('''
-    this is deprecated user run instead
-''')
-  Future<T?> execute() async {
-    Response? res;
-    await _execute().then((value) {
-      res = value;
-    }).catchError((error) {
-      this._streamError(ApiRequestError(error));
-    });
-    return responseBuilder(res);
+  Future<Either<ActionRequestError?, T?>> execute() async {
+    Response? response;
+    ActionRequestError? apiRequestError;
+    Either<ActionRequestError?, T?>? either;
+    try {
+      response = await _execute();
+      try {
+        either = right(responseBuilder(response?.data));
+        this.onSuccess(responseBuilder(response?.data));
+      } catch (e) {
+        apiRequestError = ActionRequestError(e, res: response);
+        either = left(apiRequestError);
+      }
+    } catch (e) {
+      apiRequestError = ActionRequestError(e);
+      either = left(apiRequestError);
+    }
+    if (either.isLeft() && apiRequestError != null) {
+      this.onError(apiRequestError);
+      if (ApiRequestOptions.instance!.onError != null) {
+        ApiRequestOptions.instance!.onError!(apiRequestError);
+      }
+    }
+    this.onDone();
+    return either;
   }
 
   Future<Response?> _execute() async {
@@ -153,14 +147,11 @@ abstract class RequestAction<T, R extends ApiRequest> {
         break;
     }
     _performanceUtils?.endTrack();
-    this._streamSuccess(responseBuilder(_response?.data));
+    //this._streamSuccess(responseBuilder(_response?.data));
 
     return _response;
   }
 
-  @Deprecated('''
-    this is deprecated use run instead
-''')
   void onQueue() {
     _performanceUtils?.startTrack();
     this.onStart();
@@ -181,70 +172,38 @@ abstract class RequestAction<T, R extends ApiRequest> {
     }
     _dynamicCall
         .then((value) => this._streamSuccess(responseBuilder(value?.data)))
-        .catchError((error) => this._streamError(ApiRequestError(error)))
+        .catchError((error) => this._streamError(ActionRequestError(error)))
         .then((_) => _performanceUtils?.endTrack());
   }
 
   PerformanceReport? get performanceReport => _performanceUtils?.getReport();
 
   Future<Response?> get() async {
-    try {
-      return await _requestClient?.dio.get(
-        _dynamicPath,
-        queryParameters: _dataMap,
-      );
-    } catch (e) {
-      if (e is DioError) {
-        return e.response;
-      }
-    }
-
-    return null;
+    return await _requestClient?.dio.get(
+      _dynamicPath,
+      queryParameters: _dataMap,
+    );
   }
 
   Future<Response?> post() async {
-    try {
-      return await _requestClient?.dio.post(
-        _dynamicPath,
-        data: _dataMap,
-      );
-    } catch (e) {
-      if (e is DioError) {
-        return e.response;
-      }
-    }
-
-    return null;
+    return await _requestClient?.dio.post(
+      _dynamicPath,
+      data: _dataMap,
+    );
   }
 
   Future<Response?> put() async {
-    try {
-      return await _requestClient?.dio.put(
-        _dynamicPath,
-        data: _dataMap,
-      );
-    } catch (e) {
-      if (e is DioError) {
-        return e.response;
-      }
-    }
-
-    return null;
+    return await _requestClient?.dio.put(
+      _dynamicPath,
+      data: _dataMap,
+    );
   }
 
   Future<Response?> delete() async {
-    try {
-      return await _requestClient?.dio.delete(
-        _dynamicPath,
-        data: _dataMap,
-      );
-    } catch (e) {
-      if (e is DioError) {
-        return e.response;
-      }
-    }
-
-    return null;
+    return await _requestClient?.dio.delete(
+      _dynamicPath,
+      data: _dataMap,
+    );
   }
 
   _handleRequest(R? request) {